requirements.txt
torch
torchvision
torchaudio
numpy
pandas
jupyter
opencv-python-headless
pillow
pycocotools
ultralytics
tqdm